+++ RELEASE_NOTES.md
@@ -1,8 +1,9 @@
-## Release Notes 3.0.9
+## Release Notes 3.1.0
 
 ## Changes
 
 - **VPin Mania / Table Statistics**: Added open button/open on double-click so that the table details are shown.
+- **VPin Mania / Highscores**: Added button to delete all highscores from all players from the cabinets VPin Mania account. 
 - **Preferences / Highscore Settings**: Added "NVRam Synchronization" button to re-download all available resetted nvram files. Usually you won't need this since new nvrams are downloaded on server startup, but this way you can update them e.g. in case the volume was adjusted too. The button for this can be found on the **Highscore Settings** page.
 - **iScored**: Added "Tournament Column" info to the iScored information dialog.
 - **System Manager**: Releases are now selectable via their actual name, not the tagged version anymore. Especially VPX has more meaningful names there.
@@ -21,3 +22,4 @@
 - **Launcher**: Fixed issue where multiple connections cannot be created.
 - **Preferences / Notifications**: Fixed mixed up settings for "Highscore Change" and "Highscore Scan Completed" preferences.
 - **VPin Mania / Player Ranks**: Fixed cache reset on manual reload. 
+- **PinVol**: Fixed execution of UI so that the PinVol.exe is executed always from the "resources" folder.